+++ projects/ng-wizi-bulma/src/lib/date-picker/index.ts
@@ -2,4 +2,4 @@ export * from './date-picker.component';
 export * from './date-picker-module';
 export * from './date-picker-format';
 export * from './date-picker-intl';
-export * from './date-picker-settings';
+export * from './date-picker-default-settings';